Update the component based system

+++ Runtime/Combat/HitResult.cs
@@ -4,7 +4,7 @@ namespace TransparentGames.Essentials.Combat
 {
     public struct HitResult
     {
-        public int damageDealt;
+        public float damageDealt;
         public bool wasKilled;
         public Entity hitObject;
         public bool isCritical;
